refactor: 상수를 책임에 맞는 클래스 내부로 이동한다.
